requirements.txt
wandb
pyyaml
transformers
pandas
numpy
pytorch_lightning
typing
torch
